refactor: 각 모듈 도메인별로 패키지 정리
