+++ modules/app/src/main/java/io/github/ktg/ticketing/app/api/exception/ErrorCodeHttpStatusMapper.java
@@ -1,4 +1,4 @@
-package io.github.ktg.ticketing.api.exception;
+package io.github.ktg.ticketing.app.api.exception;
 
 import io.github.ktg.ticketing.common.exception.CommonErrorCode;
 import io.github.ktg.ticketing.common.exception.ErrorCode;
