+++ modules/domain/src/main/java/io/github/ktg/ticketing/domain/payment/PaymentStatus.java
@@ -1,4 +1,4 @@
-package io.github.ktg.ticketing.domain;
+package io.github.ktg.ticketing.domain.payment;
 
 public enum PaymentStatus {
     INIT, PENDING, COMPLETE, FAILED, REFUND
