+++ modules/domain/src/main/java/io/github/ktg/ticketing/domain/reservation/ReservationStatus.java
@@ -1,4 +1,4 @@
-package io.github.ktg.ticketing.domain;
+package io.github.ktg.ticketing.domain.reservation;
 
 public enum ReservationStatus {
     WAITING_PAYMENT, EXPIRED, COMPLETED, CANCELED
