+++ modules/persistence/src/main/java/io/github/ktg/ticketing/persistence/payment/entity/PaymentJpaEntity.java
@@ -1,6 +1,7 @@
-package io.github.ktg.ticketing.entity;
+package io.github.ktg.ticketing.persistence.payment.entity;
 
-import io.github.ktg.ticketing.domain.PaymentStatus;
+import io.github.ktg.ticketing.domain.payment.PaymentStatus;
+import io.github.ktg.ticketing.persistence.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -12,18 +13,21 @@ import jakarta.persistence.Table;
 import java.time.LocalDateTime;
 import lombok.Getter;
 
-@Table(name = "payment_histories")
+@Table(name = "payments")
 @Entity
 @Getter
-public class PaymentHistoryJpaEntity {
+public class PaymentJpaEntity extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
 
-    @Column(name = "payment_id", nullable = false)
-    private Long paymentId;
+    @Column(name = "reservation_id", nullable = false)
+    private Long reservationId;
+
+    @Column(name = "user_id", nullable = false)
+    private String userId;
 
     @Column(name = "amount", nullable = false)
     private int amount;
@@ -35,6 +39,4 @@ public class PaymentHistoryJpaEntity {
     @Column(name = "payment_at", nullable = false)
     private LocalDateTime paymentAt;
 
-    @Column(name = "fail_reason")
-    private String failReason;
 }
