+++ modules/persistence/src/main/java/io/github/ktg/ticketing/persistence/reservation/entity/ReservationJpaEntity.java
@@ -1,6 +1,7 @@
-package io.github.ktg.ticketing.entity;
+package io.github.ktg.ticketing.persistence.reservation.entity;
 
-import io.github.ktg.ticketing.domain.ReservationStatus;
+import io.github.ktg.ticketing.domain.reservation.ReservationStatus;
+import io.github.ktg.ticketing.persistence.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
